requirements.txt
Django==3.1.4
django-debug-toolbar==3.2
django-extensions==3.1.0
django-filter==2.4.0
django-mptt==0.11.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
gunicorn==20.0.4
huey==2.3.2
persiantools==1.5.1
Pillow==8.0.1
psycopg2-binary==2.8.6
redis==3.5.3
